Add files via upload
This is a the final project for ISE 437 at NCSU data analytics for Industrial Engineers in python. This code was a collaborative effort between myself, Bill Lamm, Steven Godwin, Luke Hepburn and Jackson Harwell. Our groups code was among the top performing in the class. 

With this project the goal was to take data given by the professor and use it to predict 7 days of energy usage for a certain undisclosed region. 

Utilizing historical weather data, the last years peak demand and peak temperature per day we were tasked with using multitudes of data analytical practice to determine the best forecasted demand.